+++ packages/learn-card-bridge-http/CHANGELOG.md
@@ -5,4 +5,5 @@
 ### Patch Changes
 
 -   Updated dependencies
-    -   @learncard/core@1.1.5
+    -   @learncard/core@1.2.0
+    -   @learncard/types@2.1.2